+++ tools/x-env-common.sh
@@ -1,3 +1,5 @@
+# devela/tools/x-env-common.sh
+
 export RUSTFLAGS=""
 
 export RUSTDOCFLAGS="--html-in-header ./src/_doc/header.html"