+++ tools/x-env-native.sh
@@ -1,3 +1,5 @@
+# devela/tools/x-env-native.sh
+
 RUSTFLAGS="${RUSTFLAGS:+$RUSTFLAGS }-C target-cpu=native"
 RUSTDOCFLAGS="${RUSTDOCFLAGS:+$RUSTDOCFLAGS }-C target-cpu=native"