+++ tools/x-env-nightly.sh
@@ -1,3 +1,5 @@
+# devela/tools/x-env-nightly.sh
+#
 # WAIT: https://github.com/rust-lang/cargo/issues/331#issuecomment-1081704033
 # "-Zbuild-analysis", # MAYBE